test: move version sync file reader to test code
